+++ conf/script/src/build_system/cmd/setup/__main__.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
 
-import build_system.cmd.setup.cli
+import build_system.cmd.setup.cli_impl
 import utils.cli.main
 
 
 def main():
-    build_system.cmd.setup.cli.setup()
+    build_system.cmd.setup.cli_impl.setup()
 
 
 utils.cli.main.wrap_main(main)
